refactor(agent): move tool-classifier to lib/agent (shared with agent-ops)
